+++ main/threading_indexes.h
@@ -6,9 +6,9 @@
 
 using namespace std;
 
-using num_cards_t        = unsigned int;
-using num_threads_t      = unsigned int;
-using start_card_index_t = unsigned int;
+using num_cards_t        = unsigned long long int;
+using num_threads_t      = unsigned long long int;
+using start_card_index_t = unsigned long long int;
 
 using start_card_indexes_t = vector<start_card_index_t>;
